fix(core): eliminate import cycles
This change eliminates two import cycles found in `@actions/core`.
These import cycles would prevent rollupjs from removing unused code
within `@actions/core`, bloating the resultant packaged js. Anecdotally,
with these changes, a project I work on that only uses core's `info` and
`setState` was reduced in size from 1.0MiB to approximately 244KiB.

- `core.ts` <-> `oidc-utils.ts`:
  - `core.ts` importing from `oidc-utils.ts` for the purpose of
    forming and exporting `getIDToken()`.
  - `oidc-utils.ts` was importing `debug()` and `setSecret` from
    `core.ts`.
- `core.ts` <-> `utils.ts`:
  - `core.ts` importing `toCommandProperties` and `toCommandValue` from
    `utils.ts`.
  - `utils.ts` importing `AnnotationProperties` from `core.ts`.

I resolved these issues by:

- Moving the definition of `AnnotationProperties` from `core.ts` into
    `annotation.ts`, and having `utils.ts` import from there.
- Moving `getIDToken` from `core.ts` into `oidc-utils.ts`.
- Moving `core.ts` -> `api.ts`
- Moving all package exports back into `core.ts` from `api.ts`, including the export
of `getIDToken`.

Partially fixes #1436, where one of the import cycles were first
described.
