+++ packages/core/src/core.ts
@@ -1,0 +1,34 @@
+/**
+ * Core api exports
+ */
+export * from './api.js'
+
+/**
+ * Annotation types exports
+ */
+export {AnnotationProperties} from './annotation.js'
+
+/**
+ * Summary exports
+ */
+export {summary} from './summary.js'
+
+/**
+ * @deprecated use core.summary
+ */
+export {markdownSummary} from './summary.js'
+
+/**
+ * Path exports
+ */
+export {toPosixPath, toWin32Path, toPlatformPath} from './path-utils.js'
+
+/**
+ * Platform utilities exports
+ */
+export * as platform from './platform.js'
+
+/**
+ * OIDC utilities exports
+ */
+export {getIDToken} from './oidc-utils.js'
